requirements.txt
streamlit==1.22.0
python-doctr[torch]==0.6.0
easyocr==1.7.0
pillow==9.5.0
opencv-python-headless==4.7.0.72
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
pyyaml==6.0.1